requirements.txt
absl-py==1.0.0
attrs==21.4.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cycler==0.11.0
dostoevsky==0.6.0
fasttext==0.9.2
fonttools==4.31.2
geojson==2.5.0
kiwisolver==1.4.0
libretranslatepy==2.1.1
lxml==4.9.2
matplotlib==3.5.1
mediapipe==0.8.9.1
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.0.1
pip==22.0.4
protobuf==3.19.4
pybind11==2.10.1
pyowm==3.3.0
pyparsing==3.0.7
PySocks==1.7.1
python-dateutil==2.8.2
razdel==0.5.0
requests==2.28.1
setuptools==57.4.0
six==1.16.0
telebot==0.0.4
translate==3.6.1
urllib3==1.26.13
